feat(shared): introduce shared crate with api types, crypto, and telemetry
Migrate common code from `cli` and `main-app` into the new
`vulcanum-shared` crate to eliminate duplication and centralize
cross-cutting concerns. Deduplicated API request/response types
(`ConnectRequest`, `ConnectResponse`, `RefreshRequest`,
`RefreshResponse`, `StatusResponse`, `PollResponse`, `JobResponse`,
`SubmitResultRequest`), the `generate_alphanumeric_string` token helper,
and the `tracing` telemetry initialization. Updated both workspace
members to depend on `vulcanum-shared` and removed now-redundant local
definitions. Cleaned up stale `#[allow(dead_code)]` attributes and
switched `actix-rt` usage in `main-app` to the standard `actix-web`
runtime.

+++ cli/src/commands/daemon/job.rs
@@ -86,8 +86,8 @@ pub(crate) async fn handle_job(
     let result = SubmitResultRequest {
         pr_url: harness_result.pr_url.unwrap_or_default(),
         exit_code: harness_result.exit_code,
-        tokens_used: harness_result.tokens_used,
-        duration_ms: harness_result.duration_ms,
+        tokens_used: harness_result.tokens_used as i64,
+        duration_ms: harness_result.duration_ms as i64,
     };
 
     if let Err(e) = client
